Logic for converting to JSON moved to new module
